+++ src/main/java/org/jitsi/videobridge/health/config/HealthIntervalProperty.java
@@ -17,6 +17,7 @@
 package org.jitsi.videobridge.health.config;
 
 import org.jitsi.videobridge.util.config.*;
+import org.jitsi.videobridge.util.config.retriever.*;
 
 import java.util.concurrent.*;
