+++ src/main/java/org/jitsi/videobridge/octo/config/OctoRegionProperty.java
@@ -17,8 +17,8 @@
 package org.jitsi.videobridge.octo.config;
 
 import com.typesafe.config.*;
-import org.jitsi.videobridge.util.*;
 import org.jitsi.videobridge.util.config.*;
+import org.jitsi.videobridge.util.config.retriever.*;
 
 /**
  * A singleton property representing the octo's region
