+++ src/main/java/org/jitsi/videobridge/rest/WebSocketConfig.java
@@ -18,6 +18,7 @@ package org.jitsi.videobridge.rest;
 
 import com.typesafe.config.*;
 import org.jitsi.videobridge.util.config.*;
+import org.jitsi.videobridge.util.config.retriever.*;
 
 public class WebSocketConfig
 {
